requirements.txt
# Core dependencies
smolagents>=0.1.0
transformers>=4.35.0
torch>=2.0.0

# Data & Finance
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0

# API & HTTP
requests>=2.31.0
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Visualization (optional)
matplotlib>=3.7.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
ipykernel>=6.25.0
